requirements.txt
PyYAML>=6.0
accelerate>=0.29
numpy>=1.26,<2.3
torch>=2.2
transformers>=4.40
tqdm>=4.66
